Додано найчастіший хід
